# Use sync.Pool for small pages in db.allocate