@funny-falcon 2016-04-22
Closed my #409 in favor of this pull request.