@benbjohnson 2016-04-22
Sorry for the delay. I did some testing on this and it looks good. I moved the `sync.Pool` into the `DB` (https://github.com/boltdb/bolt/commit/f5f0f7af77195e5339bff58ea4346aa5491835c1) to avoid a global variable. Also, if users are splitting out `DB`s for security reasons then they probably wouldn't want pages shared between DBs in case there's a bug.

Thanks for the PR!