@funny-falcon 2016-04-22
Imho, it is better to keep global var, and ensure there is no bug.

Doubdfuly, boltdb will work without zeroing this pages before returning to pool, so bug probability is miserable, it will be always caught by tests.

But if DB were split for performance, then global pool is big gain.