@benbjohnson 2016-04-22
@funny-falcon I agree that it's unlikely that there's a bug since we're zeroing out. The `DB` is an isolated unit though and I'd like to keep that isolation. Also, I would think that you would still see a good performance gain if you're running multiple DBs with their own pools.